+++ .gitignore
@@ -143,6 +143,7 @@ data/failed/
 data/trash/
 data/gtfs/
 data/tmp
+data/agencyconf/**
 
 #these files are under app/static but they get copied to the outside directory on startup
 logging.conf
